Factor out l.lib.config.
