+++ landscape/package/releaseupgrader.py
@@ -8,7 +8,7 @@ import tarfile
 
 from twisted.internet.defer import succeed
 
-from landscape.deployment import get_bindir
+from landscape.lib.config import get_bindir
 from landscape.lib.fetch import url_to_filename, fetch_to_files
 from landscape.lib.lsb_release import parse_lsb_release, LSB_RELEASE_FILENAME
 from landscape.lib.gpg import gpg_verify
